Removed chartkick. Starting setting up data for treechart
Removed chartkick. Chartkick has limited support for configuration and advanced usage of charts. It also doesn't have an interface for the treechart. Google Charts is straightforward enough, and the documentation is complete, so I'm planning on handrolling it see

Started setting up data for treechart. The treechart requires a parent/child relationship. I started by removing an extra branch in the index route of the home_controller. Then I grabbed an appropriate field to group by ("Theme" which corresponds to category). Still have to group the data, generate summary statistics, and create the treechart server side.